+++ .swiftpm/xcode/xcshareddata/xcschemes/Example.xcscheme
@@ -50,13 +50,6 @@
             ReferencedContainer = "container:">
          </BuildableReference>
       </BuildableProductRunnable>
-      <EnvironmentVariables>
-         <EnvironmentVariable
-            key = "IIDadataAPIToken"
-            value = "abadf779d0525bebb9e16b72a97eabf4f7143292"
-            isEnabled = "YES">
-         </EnvironmentVariable>
-      </EnvironmentVariables>
    </LaunchAction>
    <ProfileAction
       buildConfiguration = "Release"
