+++ .swiftpm/xcode/xcshareddata/xcschemes/IIDadataUI.xcscheme
@@ -15,20 +15,6 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "IIDadata"
-               BuildableName = "IIDadata"
-               BlueprintName = "IIDadata"
-               ReferencedContainer = "container:">
-            </BuildableReference>
-         </BuildActionEntry>
-         <BuildActionEntry
-            buildForTesting = "YES"
-            buildForRunning = "YES"
-            buildForProfiling = "YES"
-            buildForArchiving = "YES"
-            buildForAnalyzing = "YES">
-            <BuildableReference
-               BuildableIdentifier = "primary"
                BlueprintIdentifier = "IIDadataUI"
                BuildableName = "IIDadataUI"
                BlueprintName = "IIDadataUI"
@@ -54,9 +40,18 @@
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
+      <MacroExpansion>
+         <BuildableReference
+            BuildableIdentifier = "primary"
+            BlueprintIdentifier = "IIDadataUI"
+            BuildableName = "IIDadataUI"
+            BlueprintName = "IIDadataUI"
+            ReferencedContainer = "container:">
+         </BuildableReference>
+      </MacroExpansion>
       <EnvironmentVariables>
          <EnvironmentVariable
-            key = "IIDadataAPIToken"
+            key = "IIDadataAPIkey"
             value = "abadf779d0525bebb9e16b72a97eabf4f7143292"
             isEnabled = "YES">
          </EnvironmentVariable>
