+++ Package.swift
@@ -4,51 +4,49 @@
 import PackageDescription
 
 let package = Package(
-  name: "IIDadata",
-  defaultLocalization: "ru",
-  platforms: [
-    .iOS(.v14),
-    .macOS(.v11),
-    .tvOS(.v14),
-    .watchOS(.v7),
-  ],
-  products: [
-    // Products define the executables and libraries produced by a package, and make them visible to other packages.
-    .library(
-      name: "IIDadata",
-      targets: ["IIDadata"]
-    ),
-    .library(
-      name: "IIDadataUI",
-      targets: ["IIDadata", "IIDadataUI"]
-    )
-  ],
-  dependencies: [
-    // Dependencies declare other packages that this package depends on.
-    // .package(url: /* package url */, from: "1.0.0"),
-  ],
-  targets: [
-    // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-    // Targets can depend on other targets in this package, and on products in packages which this package depends on.
-    .executableTarget(
-      name: "Example",
-      dependencies: ["IIDadata", "IIDadataUI"],
-      path: "IIDadata/Example/"
-    ),
-    .target(
-      name: "IIDadata",
-      dependencies: [],
-      path: "IIDadata/Sources/"
-    ),
-    .target(
-      name: "IIDadataUI",
-      dependencies: ["IIDadata"],
-      path: "IIDadata/Suggestions Popover/"
-    ),
-    .testTarget(
-      name: "IIDadataTests",
-      dependencies: ["IIDadata"],
-      path: "IIDadata/Tests/IIDadataTests/"
-    ),
-  ]
+	name: "IIDadata",
+	defaultLocalization: "ru",
+	platforms: [
+		.iOS(.v14),
+		.macOS(.v11),
+		.tvOS(.v14),
+		.watchOS(.v7),
+	],
+	products: [
+		// Products define the executables and libraries produced by a package, and make them visible to other packages.
+		.library(
+			name: "IIDadata",
+			targets: ["IIDadata"]
+		),
+		.library(
+			name: "IIDadataUI",
+			targets: ["IIDadata", "IIDadataUI"]
+		),
+	],
+	dependencies: [
+		// Dependencies declare other packages that this package depends on.
+		// .package(url: /* package url */, from: "1.0.0"),
+	],
+	targets: [
+		// Targets are the basic building blocks of a package. A target can define a module or a test suite.
+		// Targets can depend on other targets in this package, and on products in packages which this package depends on.
+		.executableTarget(
+			name: "Example",
+			dependencies: ["IIDadata", "IIDadataUI"],
+			path: "IIDadata/Example/"
+		),
+		.target(
+			name: "IIDadata",
+			dependencies: []
+		),
+		.target(
+			name: "IIDadataUI",
+			dependencies: ["IIDadata"]
+		),
+		.testTarget(
+			name: "IIDadataTests",
+			dependencies: ["IIDadata"],
+			path: "IIDadata/Tests/IIDadataTests/"
+		),
+	]
 )
