+++ Sources/IIDadata/DadataQueryProtocol.swift
