+++ Sources/IIDadata/Model/Address/AddressSuggestionResponse.swift
