+++ Sources/IIDadata/Model/SuggestionProtocol.swift
